Correct empty file

+++ rebdev.pavel/S2/main.cpp
@@ -45,10 +45,13 @@ int main(int argv, char ** argc)
       return 1;
     }
   }
-  for (size_t i = 0; i < (resStack.size() - 1); ++i)
+  if (!resStack.empty())
   {
-    std::cout << *(resStack.drop()) << ' ';
+    for (size_t i = 0; i < (resStack.size() - 1); ++i)
+    {
+      std::cout << *(resStack.drop()) << ' ';
+    }
+    std::cout << *(resStack.drop()) << '\n';
   }
-  std::cout << *(resStack.drop()) << '\n';
   return 0;
 }
